refactor to BIP45
