+++ projects/subgraph-bean/src/utils/LockedBeans.ts
@@ -15,9 +15,9 @@ import { loadOrCreatePool } from "./Pool";
 import { loadOrCreateTwaOracle } from "./price/TwaOracle";
 
 export function calcLockedBeans(blockNumber: BigInt): BigInt {
-  // If BIP44 is deployed - return the result from the contract
+  // If BIP45 is deployed - return the result from the contract
   // Future improvement when actual deployment block is known, can check block and avoid this call in earlier blocks.
-  if (blockNumber > BigInt.fromU32(19764699)) {
+  if (blockNumber > BigInt.fromU32(19922925)) {
     // If we are trying to calculate locked beans on the same block as the sunrise, use the values from the previous hour
     const twaOracle = loadOrCreateTwaOracle(getUnderlyingUnripe(blockNumber).toHexString());
     const twaReserves =
