+++ projects/subgraph-beanstalk/tests/event-mocking/SeedGauge.ts
@@ -9,7 +9,7 @@ import {
   UpdateGaugeSettings,
   TotalGerminatingStalkChanged,
   TotalStalkChangedFromGermination
-} from "../../generated/BIP44-SeedGauge/Beanstalk";
+} from "../../generated/BIP45-SeedGauge/Beanstalk";
 
 import { BEANSTALK } from "../../../subgraph-core/utils/Constants";
